Update to keep from rendering ListOfProjectsStore when list of 'active' projects change.
